+++ bitmap/goBitmap.h
@@ -12,7 +12,7 @@
 #include "bitmap_find_c.h"
 #include "../base/color_find_c.h"
 // #include "../screen/screen_c.h"
-#include "../base/io_c.h"
+#include "../base/file_io_c.h"
 #include "../base/pasteboard_c.h"
 #include "../base/str_io_c.h"
 #include <assert.h>
@@ -95,25 +95,6 @@ int count_of_bitmap(MMBitmapRef bitmap, MMBitmapRef sbit, float tolerance){
 	MMRect rect = MMBitmapGetBounds(bitmap);
 
 	return countOfBitmapInRect(bitmap, sbit, rect, tolerance);
-}
-
-MMPoint aFindBitmap(MMBitmapRef bit_map, MMRect rect){
-	// MMRect rect;
-	// rect.size.width = 10;
-	// rect.size.height = 20;
-	// rect.origin.x = 10;
-	// rect.origin.y = 20;
-
-	float tolerance = 0.0f;
-	MMPoint point;
-
-	tolerance = 0.5;
-
-	if (findBitmapInRect(bit_map, bit_map, &point,
-	                     rect, tolerance) == 0) {
-		return point;
-	}
-	return point;
 }
 
 bool point_in_bounds(MMBitmapRef bitmap, MMPoint point){