+++ key/goKey.h
@@ -25,6 +25,7 @@ struct KeyNames{
 	{ "delete",         K_DELETE },
 	{ "enter",          K_RETURN },
 	{ "tab",            K_TAB },
+	{ "esc",            K_ESCAPE },
 	{ "escape",         K_ESCAPE },
 	{ "up",             K_UP },
 	{ "down",           K_DOWN },
@@ -34,6 +35,7 @@ struct KeyNames{
 	{ "end",            K_END },
 	{ "pageup",         K_PAGEUP },
 	{ "pagedown",       K_PAGEDOWN },
+	//
 	{ "f1",             K_F1 },
 	{ "f2",             K_F2 },
 	{ "f3",             K_F3 },
@@ -58,12 +60,25 @@ struct KeyNames{
 	{ "f22",            K_F22 },
 	{ "f23",            K_F23 },
 	{ "f24",            K_F24 },
+	//
+	{ "cmd",            K_META },
+	{ "lcmd",           K_LMETA },
+	{ "rcmd",           K_RMETA },
 	{ "command",        K_META },
 	{ "alt",            K_ALT },
+	{ "lalt",           K_LALT },
+	{ "ralt",           K_RALT },
+	{ "ctrl",           K_CONTROL },
+	{ "lctrl",          K_LCONTROL },
+	{ "rctrl",          K_RCONTROL },
 	{ "control",        K_CONTROL },
 	{ "shift",          K_SHIFT },
-	{ "right_shift",    K_RIGHTSHIFT },
+	{ "lshift",         K_LSHIFT },
+	{ "rshift",         K_RSHIFT },
+	{ "right_shift",    K_RSHIFT },
+	{ "capslock",       K_CAPSLOCK },
 	{ "space",          K_SPACE },
+	{ "print",          K_PRINTSCREEN },
 	{ "printscreen",    K_PRINTSCREEN },
 	{ "insert",         K_INSERT },
 	{ "menu",           K_MENU },
@@ -81,6 +96,27 @@ struct KeyNames{
 	{ "audio_repeat",   K_AUDIO_REPEAT },
 	{ "audio_random",   K_AUDIO_RANDOM },
 
+	{ "num0",		K_NUMPAD_0 },
+	{ "num1",		K_NUMPAD_1 },
+	{ "num2",		K_NUMPAD_2 },
+	{ "num3",		K_NUMPAD_3 },
+	{ "num4",		K_NUMPAD_4 },
+	{ "num5",		K_NUMPAD_5 },
+	{ "num6",		K_NUMPAD_6 },
+	{ "num7",		K_NUMPAD_7 },
+	{ "num8",		K_NUMPAD_8 },
+	{ "num9",		K_NUMPAD_9 },
+	{ "num_lock",	K_NUMPAD_LOCK },
+
+	{"num.",	K_NUMPAD_DECIMAL },
+	{"num+",	K_NUMPAD_PLUS },
+	{"num-",	K_NUMPAD_MINUS },
+	{"num*",	K_NUMPAD_MUL },
+	{"num/",	K_NUMPAD_DIV },
+	{"num_clear",	K_NUMPAD_CLEAR },
+	{"num_enter",	K_NUMPAD_ENTER },
+	{"num_equal",	K_NUMPAD_EQUAL },
+
 	{ "numpad_0",		K_NUMPAD_0 },
 	{ "numpad_1",		K_NUMPAD_1 },
 	{ "numpad_2",		K_NUMPAD_2 },
@@ -91,6 +127,7 @@ struct KeyNames{
 	{ "numpad_7",		K_NUMPAD_7 },
 	{ "numpad_8",		K_NUMPAD_8 },
 	{ "numpad_9",		K_NUMPAD_9 },
+	{ "numpad_lock",	K_NUMPAD_LOCK },
 
 	{ "lights_mon_up",    K_LIGHTS_MON_UP },
 	{ "lights_mon_down",  K_LIGHTS_MON_DOWN },
@@ -130,19 +167,23 @@ int CheckKeyCodes(char* k, MMKeyCode *key){
 int CheckKeyFlags(char* f, MMKeyFlags* flags){
 	if (!flags) { return -1; }
 
-	if (strcmp(f, "alt") == 0) {
+	if ( strcmp(f, "alt") == 0 || strcmp(f, "ralt") == 0 || 
+		strcmp(f, "lalt") == 0 ) {
 		*flags = MOD_ALT;
 	}
-	else if(strcmp(f, "command") == 0) {
+	else if( strcmp(f, "command") == 0 || strcmp(f, "cmd") == 0 || 
+		strcmp(f, "rcmd") == 0 || strcmp(f, "lcmd") == 0 ) {
 		*flags = MOD_META;
 	}
-	else if(strcmp(f, "control") == 0) {
+	else if( strcmp(f, "control") == 0 || strcmp(f, "ctrl") == 0 ||
+	 strcmp(f, "rctrl") == 0 || strcmp(f, "lctrl") == 0 ) {
 		*flags = MOD_CONTROL;
 	}
-	else if(strcmp(f, "shift") == 0) {
+	else if( strcmp(f, "shift") == 0 || strcmp(f, "right_shift") == 0 || 
+		strcmp(f, "rshift") == 0 || strcmp(f, "lshift") == 0 ) {
 		*flags = MOD_SHIFT;
 	}
-	else if(strcmp(f, "none") == 0) {
+	else if( strcmp(f, "none") == 0 ) {
 		*flags = (MMKeyFlags) MOD_NONE;
 	} else {
 		return -2;
@@ -155,19 +196,20 @@ int GetFlagsFromValue(char* value[], MMKeyFlags* flags, int num){
 	if (!flags) {return -1;}
 
 	int i;
-	for ( i= 0; i <num; i++){
+	for ( i= 0; i <num; i++) {
 		MMKeyFlags f = MOD_NONE;
 		const int rv = CheckKeyFlags(value[i], &f);
-		if (rv) {return rv;}
+		if (rv) { return rv; }
 
 		*flags = (MMKeyFlags)(*flags | f);
 	}
+
 	return 0;
 	// return CheckKeyFlags(fstr, flags);
 }
 
 // If it's not an array, it should be a single string value.
-char* key_Tap(char *k, char* keyArr[], int num, int keyDelay){
+char* key_Taps(char *k, char* keyArr[], int num, int keyDelay){
 	MMKeyFlags flags = MOD_NONE;
 	// MMKeyFlags flags = 0;
 	MMKeyCode key;
@@ -194,7 +236,8 @@ char* key_Tap(char *k, char* keyArr[], int num, int keyDelay){
 			microsleep(keyDelay);
 	}
 
-	return "0";
+	// return "0";
+	return "";
 }
 
 char* key_tap(char *k, char *akey, char *keyT, int keyDelay){
@@ -214,7 +257,7 @@ char* key_tap(char *k, char *akey, char *keyT, int keyDelay){
 					return "Invalid key flag specified.";
 					break;
 			}
-		}else{
+		} else {
 			char* akeyArr[2] = {akey, keyT};
 			switch(GetFlagsFromValue(akeyArr, &flags, 2)) {
 				case -1:
@@ -239,7 +282,7 @@ char* key_tap(char *k, char *akey, char *keyT, int keyDelay){
 			microsleep(keyDelay);
 	}
 
-	return "0";
+	return "";
 }
 
 char* key_toggle(char *k, char *d, char *akey, char *keyT){
@@ -250,7 +293,7 @@ char* key_toggle(char *k, char *d, char *akey, char *keyT){
 	// char *k;
 	// k = *kstr;
 
-	if (d != 0){
+	if (d != 0) {
 		// char *d;
 		// d = *dstr;
 		if (strcmp(d, "down") == 0) {
@@ -297,7 +340,7 @@ char* key_toggle(char *k, char *d, char *akey, char *keyT){
 			microsleep(keyboardDelay);
 	}
 
-	return "0";
+	return "";
 }
 
 void type_string(char *str){